Removed all crap files from version control
Bascially visual studio generates a numbe of file when trying to work
with it on a project. Specifically temp user files and build related
files. The new gitignore does have all Windows and Visual Studio crap
removed. In contrast, the obj/ and bin/ directories are completely
removed from version control so is the *.suo file.